feat: Refactor reviews controller and routes (changed review routes to standard new show and create
This commit refactors the `ReviewsController` by separating the `new_review` action into `new` and `create` actions. It also updates the `review_params` method to permit the `activity_id` parameter. Additionally, the routes for creating and showing reviews have been updated to use the `resources` keyword. This improves the code structure and adheres to RESTful conventions.